Functional test suite added
